# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "基于 Golang 开发的高并发支付系统核心 API",
        "title": "支付系统核心 API",
        "termsOfService": "https://example.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/orders": {
            "post": {
                "description": "创建支付订单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "创建订单",
                "parameters": [
                    {
                        "description": "订单信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/orders/query": {
            "get": {
                "description": "根据商户订单号查询订单",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "查询订单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商户订单号",
                        "name": "out_order_no",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "订单不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/orders/{order_no}": {
            "get": {
                "description": "根据订单号获取订单详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "获取订单信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "订单号",
                        "name": "order_no",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "订单不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "检查服务健康状态，包括数据库和 Redis 连接状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统"
                ],
                "summary": "健康检查",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.CreateOrderRequest": {
            "type": "object",
            "required": [
                "money",
                "notify_money",
                "notify_url",
                "out_order_no"
            ],
            "properties": {
                "extra": {
                    "type": "string",
                    "example": "{}"
                },
                "jump_url": {
                    "type": "string",
                    "example": "https://example.com/jump"
                },
                "merchant_tax": {
                    "type": "integer",
                    "example": 50
                },
                "money": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 10000
                },
                "notify_money": {
                    "type": "integer",
                    "example": 10000
                },
                "notify_url": {
                    "type": "string",
                    "example": "https://example.com/notify"
                },
                "out_order_no": {
                    "type": "string",
                    "example": "ORD20240101001"
                },
                "pay_channel_id": {
                    "type": "integer",
                    "example": 1
                },
                "product_id": {
                    "type": "string",
                    "example": "PROD001"
                },
                "product_name": {
                    "type": "string",
                    "example": "支付宝扫码"
                },
                "req_extra": {
                    "type": "string",
                    "example": "{}"
                },
                "tax": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}